Temp rename /event to fix casing issue
