+++ client/src/pages/event-temp-rename/EventDetails.css
