+++ client/src/router.tsx
@@ -1,12 +1,12 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "./App";
 import BarPage from "./pages/BarPage/BarPage";
-import EventDetails from "./pages/Event/EventDetails";
-import Events from "./pages/Event/Events";
 import Home from "./pages/Home/Home";
 import Login from "./pages/Login/Login";
 import MusicGroup from "./pages/MusicGroup/MusicGroup";
 import Register from "./pages/Register/Register";
+import EventDetails from "./pages/event-temp-rename/EventDetails";
+import Events from "./pages/event-temp-rename/Events";
 
 const router = createBrowserRouter([
   {
